cut the production plan and paste in shooting schedule page in pre produ
